flux_syntax

Module lexer

Modules

token

Convenience module so we can refer to token kinds as token::*

Structs

Cursor

Frame 🔒

Token

Lit

A literal token.

Enums

TokenKind

Delimiter

Describes how a sequence of token trees is delimited. Cannot use proc_macro::Delimiter directly because this structure should implement some additional traits.

LitKind