flux_syntax::parser

Trait Peek

trait Peek {
    // Required method
    fn peek_at(self, tokens: &mut Cursor<'_>, pos: usize) -> bool;
}

A trait for testing whether a single token matches some rule. This is mainly implemented for [Token] to test whether a token is equal to a specific token.

Required Methods

fn peek_at(self, tokens: &mut Cursor<'_>, pos: usize) -> bool

Returns true if the token at the requested position matches this peek rule