flux_syntax

Module surface

Modules

visit

A visitor for types in surface

Structs

BaseTy

ConstArg

ConstantInfo

DetachedInherentImpl

DetachedItem

DetachedSpecs

DetachedTrait

DetachedTraitImpl

EnumDef

Expr

ExprPath

A Path but for refinement expressions

ExprPathSegment

FieldExpr

FnOutput

FnSig

GenericArg

GenericParam

Generics

Impl

ImplAssocReft

ImplItemFn

Indices

Item

LetDecl

NodeId

A NodeId is a unique identifier we assign to some AST nodes to be able to attach information to them. For example, to assign a resolution to a Path. The NodeId is unique within a crate.

Path

PathSegment

PrimOpProp

A (currently global) primop property; see tests/tests/pos/surface/

Punctuated

A punctuated sequence of values of type T separated by punctuation of type P

Qualifier

RefineParam

Requires

SortDecl

SortPath

A Path but for sorts.

SpecFunc

A global function definition. It can be either an uninterpreted function or a syntactic abstraction, i.e., a function with a body.

Spread

StructDef

Trait

TraitAssocReft

TraitItemFn

TraitRef

Ty

TyAlias

VariantDef

VariantRet

WhereBoundPredicate

Ident

Lit

A literal token.

Span

A compressed span.

Enums

Async

Attr

An attribute attaches metadata to an item.

BaseSort

BaseTyKind

BinOp

BindKind

ConstArgKind

ConstructorArg

DetachedItemKind

Ensures

ExprKind

FluxItem

FnInput

FnRetTy

GenericArgKind

Ignored

A boolean-like enum used to mark whether a piece of code is ignored.

ItemKind

ParamMode

QuantKind

RefineArg

Sort

Trusted

A boolean-like enum used to mark whether some code should be trusted.

TyKind

UnOp

LitKind

Mutability

Type Aliases

GenericBounds

RefineParams