Struct flux_syntax::surface::Ident

pub struct Ident {
    pub name: Symbol,
    pub span: Span,
}

Fields
name: Symbolspan: Span
Implementations

impl Ident

pub const fn new(name: Symbol, span: Span) -> Ident

Constructs a new identifier from a symbol and a span.

pub const fn with_dummy_span(name: Symbol) -> Ident

Constructs a new identifier with a dummy span.

pub fn empty() -> Ident

pub fn from_str(string: &str) -> Ident

Maps a string to an identifier with a dummy span.

pub fn from_str_and_span(string: &str, span: Span) -> Ident

Maps a string and a span to an identifier.

pub fn with_span_pos(self, span: Span) -> Ident

Replaces lo and hi with those from span, but keep hygiene context.

pub fn without_first_quote(self) -> Ident

pub fn normalize_to_macros_2_0(self) -> Ident

“Normalize” ident for use in comparisons using “item hygiene”. Identifiers with same string value become same if they came from the same macro 2.0 macro (e.g., macro item, but not macro_rules item) and stay different if they came from different macro 2.0 macros. Technically, this operation strips all non-opaque marks from ident’s syntactic context.

pub fn normalize_to_macro_rules(self) -> Ident

“Normalize” ident for use in comparisons using “local variable hygiene”. Identifiers with same string value become same if they came from the same non-transparent macro (e.g., macro or macro_rules! items) and stay different if they came from different non-transparent macros. Technically, this operation strips all transparent marks from ident’s syntactic context.

pub fn as_str(&self) -> &str

Access the underlying string. This is a slowish operation because it requires locking the symbol interner.

Note that the lifetime of the return value is a lie. See Symbol::as_str() for details.

impl Ident

pub fn is_special(self) -> bool

Returns true for reserved identifiers used internally for elided lifetimes, unnamed method parameters, crate root module, error recovery etc.

pub fn is_used_keyword(self) -> bool

Returns true if the token is a keyword used in the language.

pub fn is_unused_keyword(self) -> bool

Returns true if the token is a keyword reserved for possible future use.

pub fn is_reserved(self) -> bool

Returns true if the token is either a special identifier or a keyword.

pub fn is_path_segment_keyword(self) -> bool

A keyword or reserved identifier that can be used as a path segment.

pub fn is_raw_guess(self) -> bool

We see this identifier in a normal identifier position, like variable name or a type. How was it written originally? Did it use the raw form? Let’s try to guess.