Struct flux_syntax::surface::Punctuated

pub struct Punctuated<T, P> {
    inner: Vec<(T, P)>,
    last: Option<Box<T>>,
}

A punctuated sequence of values of type T separated by punctuation of type P

Fields
inner: Vec<(T, P)>last: Option<Box<T>>
Implementations

impl<T, P> Punctuated<T, P>

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

Determines whether this punctuated sequence is empty, meaning it contains no syntax tree nodes or punctuation.

pub fn push_value(&mut self, value: T)

Appends a syntax tree node onto the end of this punctuated sequence. The sequence must already have a trailing punctuation, or be empty.

Use push instead if the punctuated sequence may or may not already have trailing punctuation.

Panics

Panics if the sequence is nonempty and does not already have a trailing punctuation.

pub fn empty_or_trailing(&self) -> bool

Returns true if either this Punctuated is empty, or it has a trailing punctuation.

Equivalent to punctuated.is_empty() || punctuated.trailing_punct().

pub fn trailing_punct(&self) -> bool

Determines whether this punctuated sequence ends with a trailing punctuation.

pub fn into_values(self) -> Vec<T>