flux_syntax::surface

Module visit

A visitor for types in surface

In general there’s no specific order except that refinement parameters should be visited in an order that matches their scope. Name resolution relies on this gurantee.

Traits

Visitor

Functions

walk_base_sort

walk_bty

walk_constant

walk_defn

walk_ensures

walk_enum_def

walk_expr

walk_field_expr

walk_flux_item

walk_fn_input

walk_fn_output

walk_fn_ret_ty

walk_fn_sig

walk_fn_trait_ref

walk_generic_arg

walk_generic_param

walk_generics

walk_impl

walk_impl_assoc_reft

walk_impl_item

walk_indices

walk_item

walk_path

walk_path_expr

walk_path_expr_segment

walk_path_segment

walk_primop_prop

walk_qualifier

walk_refine_arg

walk_refine_param

walk_sort

walk_sort_decl

walk_sort_path

walk_struct_def

walk_trait

walk_trait_assoc_reft

walk_trait_item

walk_trait_ref

walk_ty

walk_ty_alias

walk_variant

walk_variant_ret

walk_where_predicate