Type Alias flux_syntax::LalrpopError

pub(crate) type LalrpopError = ParseError<Location, Token, UserParseError>;

Aliased Type

enum LalrpopError {
    InvalidToken {
        location: Location,
    },
    UnrecognizedEof {
        location: Location,
        expected: Vec<String, Global>,
    },
    UnrecognizedToken {
        token: (Location, Token, Location),
        expected: Vec<String, Global>,
    },
    ExtraToken {
        token: (Location, Token, Location),
    },
    User {
        error: UserParseError,
    },
}

Variants

InvalidToken

Fields

location: Location

Generated by the parser when it encounters a token (or EOF) it did not expect.

UnrecognizedEof

Fields

location: Location

The end of the final token

expected: Vec<String, Global>

The set of expected tokens: these names are taken from the grammar and hence may not necessarily be suitable for presenting to the user.

Generated by the parser when it encounters an EOF it did not expect.

UnrecognizedToken

Fields

token: (Location, Token, Location)

The unexpected token of type T with a span given by the two L values.

expected: Vec<String, Global>

The set of expected tokens: these names are taken from the grammar and hence may not necessarily be suitable for presenting to the user.

Generated by the parser when it encounters a token it did not expect.

ExtraToken

Fields

token: (Location, Token, Location)

Generated by the parser when it encounters additional, unexpected tokens.

User

Fields

error: UserParseError

Custom error type.