flux_syntax::parser::lookahead

Trait Peek

pub(crate) trait Peek: Copy {
    // Required methods
    fn matches(self, tok: Token) -> bool;
    fn display(self) -> impl Iterator<Item = &'static str>;
}

A trait for testing whether a token matches a rule.

This trait is primarily implemented for Token to test for exact equality.

Required Methods

fn matches(self, tok: Token) -> bool

Returns true if a token matches this rule

fn display(self) -> impl Iterator<Item = &'static str>

A string representation of the list of tokens matched by this rule. This is used to construct an error when using Lookahead1.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.