flux_syntax::parser::lookahead

Trait Peek

pub(crate) trait Peek: Copy {
    // Required method
    fn matches(self, tok: TokenKind, edition: Edition) -> bool;
}

A trait for testing whether a token matches a rule.

Required Methods

fn matches(self, tok: TokenKind, edition: Edition) -> bool

Returns true if a token matches this rule

The rule is edition dependent because keywords can vary.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.