Module flux_syntax::surface

Modules

visit

Structs

AliasReft

<qself as path>::name

ArrayLen

BaseTy

ConstSig

EnumDef

Expr

FnOutput

FnSig

FnSpec

GenericArg

GenericParam

Generics

Ident

Impl

ImplAssocReft

Indices

Lit

A literal token.

NodeId

A NodeId is a unique identifier we assign to some AST nodes to be able to attach information to them. For example, to assign a resolution to a Path. The NodeId is unique within a crate.

Path

PathExpr

PathSegment

QualNames

Qualifier

RefineParam

RefinedBy

SortDecl

SortPath

A Path but for sorts. Currently, we only support paths with one segment, so one can hardly call this a path, but we may change this later if we improve the resolver.

SpecFunc

A global function definition. It can be either an uninterpreted function or a syntactic abstraction, i.e., a function with a body.

StructDef

Trait

TraitAssocReft

TraitRef

Ty

TyAlias

VariantDef

VariantRet

WhereBoundPredicate

Enums

Arg

Async

BaseSort

BaseTyKind

BinOp

BindKind

Constraint

ExprKind

FnRetTy

GenericArgKind

GenericParamKind

Item

LitKind

Mutability

RefineArg

Sort

TyKind

UnOp

Type Aliases

GenericBounds